Copy each field of buffer header separately
Instead of copying the entire buffer header as one, each field is
copied separately in Use/AllocateBuffer when moving across processors.
This is necessary since different compilers on different processors
result in different sizes of the buffer header structure (extra
padding is added in some cases). This change ensures that only valid
fields are copied and any padding added by the compiler is neglected.

Signed-off-by: Aditya Monga <admonga@ti.com>