Update mark buffer implementation.
Mark buffer implememtation has been updated. Previously the
implementation was assuming all mark components to be proxies and
acting accordingly. This has been updated now. For all component
handles received for marking, GetComponentVersion is called and
based on the component name, it is decided whether it is a remote
component or not and accordingly action is taken. Also during
GetHandle, the actual remote component handle is also received
(along with the RPC context handle used normally). This is used
specifically during mark cases where we want to send the actual
component handle.

Signed-off-by: Aditya Monga <admonga@ti.com>
Signed-off-by: Daniel Diaz <ddiaz@ti.com>