rpc/proxy: Send buffer pointer to remote core in FreeBuffer.
The pBuffer pointer is now being passed explicitly to the remote core
during FreeBuffer. The remote core still picks up the buffer pointer from
its local header for use, however, if a NULL buffer is passed by the client
then the remote core can pick this up now and decide whether to return error
for this. Passing a NULL buffer is allowed if NPA mode is enabled but will
result in an error otherwise.

Signed-off-by: Aditya Monga <admonga@ti.com>
Signed-off-by: Daniel Diaz <ddiaz@ti.com>