proxy: Set mark info to NULL in EBD callback.
In case the local buffer header on A9 has some old mark info, it should
be set to NULL so as to not confuse the client. In EBD callback,
mark info will anyway be NULL. Component is expected to clear the mark
info and send mark callback (if it is target) or propagate the mark
info to the output buffer (if it is not target).

Signed-off-by: Aditya Monga <admonga@ti.com>
Signed-off-by: Daniel Diaz <ddiaz@ti.com>