rpc/proxy: New error handling, bookkeeping of TILER mem requests.
Update the error handling path. Now clean up does not happen in the
fault handler. The fault handler simply notifies the client. The
clean up occurs when the client call FreeHandle after receiving
critical error notification from the fault handler.

Also bookkeeping of all TILER memory requests from Ducati is done so
that in case of critical faults, this memory can be cleaned up. The
clean up is done in the last FreeHandle call from that process at the
same time when RCM and IPC are cleaned up. In case user does not tear
down properly by calling FreeHandle and simply exits the process,
Syslink takes care of the cleanup.

Signed-off-by: Aditya Monga <admonga@ti.com>
Signed-off-by: Daniel Diaz <ddiaz@ti.com>